+++ examples/timed.rs
@@ -4,19 +4,22 @@ use std::time::Instant;
 
 use ocl::{ProQue, SimpleDims, Buffer};
 
+const DATASET_SIZE: usize = 5000000;
 const SCALAR: f32 = 1.0;
-const KERNEL_ITERS: i32 = 3;
+const KERNEL_ITERS: i32 = 1;
 const BUFFER_READ_ITERS: i32 = 1;
+const PRINT_SOME_RESULTS: bool = false;
 
 fn main() {
     // Define a kernel:
     let src = r#"
-        __kernel void multiply(
-                    __global float* buffer, 
-                    __private float scalar) 
+        __kernel void add(
+                    __global float const* const source, 
+                    __private float scalar,
+                    __global float* const result) 
         {
             uint idx = get_global_id(0);
-            buffer[idx] += scalar;
+            result[idx] = source[idx] + scalar;
         }
     "#;
 
@@ -24,20 +27,21 @@ fn main() {
     let ocl_pq = ProQue::builder().src(src).build().unwrap();
 
     // Set our work dimensions / data set size to something arbitrary:
-    let dims = SimpleDims::One(5000000);
+    let dims = SimpleDims::One(DATASET_SIZE);
 
-    // Create a 'Buffer' with a built-in vector and initialize it with random 
-    // floats between 0.0 and 20.0:
-    let mut buffer: Buffer<f32> = Buffer::with_vec_scrambled(
+    // Create init and result buffers:
+    let buffer_init: Buffer<f32> = Buffer::with_vec_scrambled(
          (100.0, 200.0), &dims, &ocl_pq.queue());
+    let mut buffer_result: Buffer<f32> = Buffer::with_vec(&dims, &ocl_pq.queue());
 
     // Create a kernel with arguments matching those in the kernel:
-    let kern = ocl_pq.create_kernel("multiply", dims.work_dims()).unwrap()
-        .arg_buf(&buffer)
-        .arg_scl(SCALAR);
+    let mut kern = ocl_pq.create_kernel("add", dims.work_dims()).unwrap()
+        .arg_buf_named("source", Some(&buffer_init))
+        .arg_scl(SCALAR)
+        .arg_buf(&buffer_result);
 
-    // Keep track of inital values:
-    let init_vals = buffer.vec().unwrap().clone();
+    // // Keep track of inital values:
+    // let init_vals = buffer.vec().unwrap().clone();
 
     // #################################
     // ############ KERNEL #############
@@ -46,8 +50,14 @@ fn main() {
     // Start kernel timer
     let kern_start = Instant::now();
 
-    // Enqueue kernel a bunch:
-    for _ in 0..KERNEL_ITERS {
+    // Enqueue kernel the first time:
+    kern.enqueue(None, None);
+
+    // Set kernel source buffer to the same as result:
+    kern.set_arg_buf_named("source", Some(&buffer_result));
+
+    // Enqueue kernel for additional iterations:
+    for i in 0..(KERNEL_ITERS - 1) {
         kern.enqueue(None, None);
     }
     
@@ -60,11 +70,10 @@ fn main() {
     // Start kernel timer
     let buffer_start = Instant::now();
 
-    // // Read results from the device into buffer's local vector:
-    // for _ in 0..BUFFER_READ_ITERS {
-    //     buffer.fill_vec();
-    // }
-    buffer.fill_vec();
+    // Read results from the device into buffer's local vector:
+    for _ in 0..BUFFER_READ_ITERS {
+        buffer_result.fill_vec();
+    }
 
     print_elapsed("Buffer Read:", buffer_start);
 
@@ -72,13 +81,21 @@ fn main() {
     // ############ VERIFY #############
     // ################################# 
 
-    for (&init, &result) in init_vals.iter().zip(buffer.iter()) {
-        let correct = init + (KERNEL_ITERS as f32 * SCALAR);
-        assert!((correct - result) < 0.0001, 
-            "init: {}, correct: {}, result: {}", init, correct, result);
-    }
+    print!("Verifying result values... ");
 
-    println!("The value at index [{}] is '{}'!", 9001, buffer[9001]);  
+    for idx in 0..DATASET_SIZE {
+        let correct = buffer_init[idx] + (KERNEL_ITERS as f32 * SCALAR);
+        // let correct = buffer_init[i] + SCALAR;
+        assert!((correct - buffer_result[idx]) < 0.001, 
+            "init: {}, correct: {}, result: {}", buffer_init[idx], correct, buffer_result[idx]);
+
+        if PRINT_SOME_RESULTS && idx < DATASET_SIZE && idx > DATASET_SIZE - 20 {
+            println!("init: {}, correct: {}, result: {}", buffer_init[idx], correct, buffer_result[idx]);
+        }
+    }
+    println!("All result values are correct.");
+
+    if PRINT_SOME_RESULTS { println!("The value at index [{}] is '{}'!", 9001, buffer_result[9001]); }
 }